Fix "Redefinition of module 'DoubleConversion'"
Add `:modular_headers => false` in `react_native_pods.rb` for pods `DoubleConversion`, `glog` and `Folly` in order to avoid the `Redefinition of module` error when building a swift project with `use_modular_headers!` set in Podfile. 

The current workaround is to copy all pods from the script and paste them in our Podfile, but this cause problems when we update our react-native version, as we must manually check the file and copy paste again.